Game-match solo con la protección del cambio de hash de SPA
En este commit se ha intentado bloquear todos las "guardas" que protegían game-math desde
tournamentUI. Para ello se ha includo el evento pagehide que controlaba el borrado de los
tempusers de la misma manera que los eventos de key y click, de fora que ahora se puede
activar o desactivar.

Además inlcuyo en el capturador de las teclas de refresco la condición de que el
window.location.hash sea el tournament-lobby dejando el resto libre

Para revertir estos cambios cambiar los archivos tournament y tournamentUI por los del
commit anterior

La guarda se activa al entrar en la selección de los jugadores y se desactiva al lanzar
la partida si es 1v1, volviendose a activar al actualizar el bracket tras handlematchresult

+++ frontend/src/js/tournament/TournamentUI.js
@@ -63,6 +63,7 @@ export class TournamentUI {
         };
         this.boundClickHandler = null;
         this.boundKeyHandler = null;
+        this.boundPageHideHandler = null;
         console.log("TournamentUI constructor - tournament:", tournament);
         this.tournament = tournament;
         // this.boundOnLeavingTournamentLobby = this.onLeavingTournamentLobby.bind(this);
@@ -174,6 +175,7 @@ export class TournamentUI {
                     }
                     console.log("InAddEevnlisteners - Preparing players for tournament with number of players: ", numberOfPlayers);
                     this.preparePlayers(numberOfPlayers);
+                    console.log("setupEventListeners: enabling hash Guard");
                     this.enableTournamentHashGuard();
                 });
                 const goBackButton = document.getElementById('tournament-back-button');
@@ -698,7 +700,9 @@ export class TournamentUI {
             // Para que afecte solo al hash tournament-lobby
             // if (((keyboardEvent.ctrlKey && (keyboardEvent.key === "F5" || keyboardEvent.key === "r")) || keyboardEvent.key === "F5")
             // 		&& window.location.hash.includes('#tournament-lobby') ){
-            if ((keyboardEvent.ctrlKey && (keyboardEvent.key === "F5" || keyboardEvent.key === "r")) || keyboardEvent.key === "F5") {
+            // Código modificado para que solo afecte en el step tournament-lobby
+            if (((keyboardEvent.ctrlKey && (keyboardEvent.key === "F5" || keyboardEvent.key === "r")) || keyboardEvent.key === "F5")
+                && window.location.hash.includes('#tournament-lobby')) {
                 keyboardEvent.preventDefault();
                 keyboardEvent.stopPropagation();
                 keyboardEvent.stopImmediatePropagation();
@@ -714,10 +718,33 @@ export class TournamentUI {
                     // if (confirmExit && (keyboardEvent.key === "F5" || keyboardEvent.key === "r")) {
                     // 	//TODO: Pedro Aquí se podría incluir lo que quisieras pero los usuarios temporales y el torneo 
                     // ya se han reseteado..
-                    // 	location.reload();
+                    // location.reload();
                     // }
                 }
             }
+            ///// 		Inicio de código comentado para probar nnuevas formulas para el refreco de game-match
+            //
+            // if ((keyboardEvent.ctrlKey && (keyboardEvent.key === "F5" || keyboardEvent.key === "r")) || keyboardEvent.key === "F5"){	
+            // 	keyboardEvent.preventDefault();
+            // 	keyboardEvent.stopPropagation();
+            // 	keyboardEvent.stopImmediatePropagation();
+            // 	if (window.location.hash.includes('#tournament-lobby') ){
+            // 		const confirmExit = confirm("Are you sure you want to reload and reset the tournament?");
+            // 		if (confirmExit && (keyboardEvent.key === "F5" || keyboardEvent.key === "r")) {
+            // 			this.resetTournament();
+            // 			location.reload();
+            // 		}
+            // 	}
+            // 	if (window.location.hash.includes('#game-match') ){
+            // 		// const confirmExit = confirm("Are you sure you want to reload and reset the game?");
+            // 		// if (confirmExit && (keyboardEvent.key === "F5" || keyboardEvent.key === "r")) {
+            // 		// 	//TODO: Pedro Aquí se podría incluir lo que quisieras pero los usuarios temporales y el torneo 
+            // 		// ya se han reseteado..
+            // 		// location.reload();
+            // 		// }
+            // 	}
+            // }
+            ////       Fin de codigo comentado para probar nuevas formulas para controlar el refrescar la página en Game-match
             if (keyboardEvent.key === "Escape") {
                 keyboardEvent.preventDefault();
                 const confirmExit = confirm("this will lead yo to Home. Are you sure you want to exit the tournament?");
@@ -729,20 +756,32 @@ export class TournamentUI {
             }
         }
     }
+    /**
+     * pagehide event is triggered as beforeonload event but is more reliable than beforeunload
+     * when a fetch call is included.
+     * This event does not show a message.
+     */
     enableTournamentHashGuard() {
-        window.addEventListener("pagehide", () => {
-            const tournamentId = this.tournament.getTournamentId();
-            if (tournamentId !== null && tournamentId !== undefined) {
-                fetch('https://localhost:8443/back/delete_user_by_tournament_id', {
-                    method: 'DELETE',
-                    headers: {
-                        "Content-Type": "application/json",
-                    },
-                    body: JSON.stringify({ TournamentId: tournamentId.toString() }),
-                    keepalive: true
-                });
-            }
-        });
+        console.log("enableTournamentHashGuard: enabled.");
+        if (!this.boundPageHideHandler) {
+            this.boundPageHideHandler = () => {
+                console.log("Page is being hidden or unloaded.");
+                const tournamentId = this.tournament.getTournamentId();
+                if (tournamentId !== null && tournamentId !== undefined) {
+                    console.log("enableTournamentHashGuard: Deleting temp users for tournamentId on pagehide:", tournamentId);
+                    fetch('https://localhost:8443/back/delete_user_by_tournament_id', {
+                        method: 'DELETE',
+                        headers: {
+                            "Content-Type": "application/json",
+                        },
+                        body: JSON.stringify({ TournamentId: tournamentId.toString() }),
+                        keepalive: true
+                    });
+                }
+            };
+            console.log("addEventListener(\"pagehide\")");
+            window.addEventListener("pagehide", this.boundPageHideHandler);
+        }
         if (!this.boundClickHandler) {
             this.boundClickHandler = this.evaluarMovimiento.bind(this);
             document.addEventListener('click', this.boundClickHandler, true); // `true` para capturar antes del default
@@ -753,6 +792,7 @@ export class TournamentUI {
         }
     }
     disableTournamentHashGuard() {
+        console.log("Tournament Hash Guard enabled.");
         if (this.boundClickHandler) {
             document.removeEventListener('click', this.boundClickHandler, true);
             this.boundClickHandler = null;
@@ -760,6 +800,11 @@ export class TournamentUI {
         if (this.boundKeyHandler) {
             document.removeEventListener('keydown', this.boundKeyHandler, true);
             this.boundKeyHandler = null;
+        }
+        if (this.boundPageHideHandler) {
+            console.log("removeEventListener(\"pagehide\")");
+            window.removeEventListener("pagehide", this.boundPageHideHandler);
+            this.boundPageHideHandler = null;
         }
     }
 }
